# Also calling watchFile on collect, so that initially cached files get watched.